+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed an issue where changing "Auto-open a file on launch" settings caused unintended navigation.
+- Fixed an issue where the calculator's last sync time was not persisted across application restarts.
 
 ## [3.5.0] - 2026-04-09
 ### Added